Admin bookings updates
